backend/requirements.txt
flask
flask-cors
numpy
pillow
matplotlib
torch
torchvision
tqdm
scipy
scikit-learn
flask-cors